chore: moved user detail page into pages folder
