+++ packages/frontend/src/App.tsx
@@ -5,8 +5,8 @@ import { Footer } from './components/Footer';
 import { Menu } from './components/Menu';
 import HomePage from './pages/Home';
 import { SponsorPage } from './pages/SponsorPage';
+import UserDetail from './pages/UserDetail';
 import SeatSelectionPage from './pages/seat_selection';
-import UserDetail from './UserDetail';
 
 const App: React.FC = () => {
   return (
